Fix requested
Display the doctor's name instead of the doctor's specialty.
Display the patient's full name instead of just the last name.
Display the complete history of all patients to the administrator.
Display of information as text instead of a button or link.
Access button to the profile of a doctor or patient added, according to the role assigned to the user